requirements.txt
Django>=5.1
psycopg2-binary
gunicorn
celery
redis
djangorestframework
djangorestframework-simplejwt
fake_useragent
requests
bs4
selenium>=4.19.0
webdriver-manager>=4.0.1
lxml
python-dotenv
django-filter>=25.1
django-cors-headers